Moved the movement system into a separate folder, clean separation by amount of possible axis.
