--- Updatable.h
@@ -1,17 +1,0 @@
-/*
- * Updatable.h
- *
- *  Created on: Jan 3, 2016
- *      Author: xasin
- */
-
-#ifndef CODE_UPDATABLE_H_
-#define CODE_UPDATABLE_H_
-
-class Updatable {
-public:
-	virtual void update() {};
-};
-
-
-#endif /* CODE_UPDATABLE_H_ */
